Criando mapeamento e service

+++ BlogFlow.Infrastructure/Repositories/UsuarioRepository.cs
@@ -50,12 +50,14 @@ namespace BlogFlow.Infrastructure.Repositories
             if (usuario == null)
                 return false;
 
+            if(usuario.Postagens != null && usuario.Postagens.Any())
+
             _dbContext.Usuarios.Remove(usuario);
             await _dbContext.SaveChangesAsync();
             return true;
         }
 
-        public async Task<IEnumerable<Usuario>> Listar(int usuarioId)
+        public async Task<IEnumerable<Usuario>> Listar()
         {
             return await _dbContext.Usuarios.Include(u => u.Postagens).ToListAsync();
         }
